+++ src/app/page.tsx
@@ -3,6 +3,8 @@ import ServiceFeatures from "@/components/home/ServiceFeatures";
 import CategorySection from "@/components/home/CategorySection";
 import FeaturedProducts from "@/components/home/FeaturedProducts";
 import FlashSaleSection from "@/components/home/FlashSaleSection";
+import ProductDetailsPreview from "@/components/product/ProductDetailsPreview";
+import NewsletterSection from "@/components/home/NewsletterSection";
 
 export default function Home() {
   return (
@@ -12,6 +14,8 @@ export default function Home() {
       <CategorySection />
       <FeaturedProducts />
       <FlashSaleSection />
+      <ProductDetailsPreview />
+      <NewsletterSection />
     </>
   );
 }